# docs/swagger.yaml
openapi: 3.0.3
info:
    title: Recipe API - OpenAPI 3.0
    description: |-
        This documentation includes **endpoints**, how to authenticate, and all the information you need to know about the API

         # Overview

         1.  The API has an access <a href="#rate-limits">rate limit</a> applied to it
         2.  Response to every request is sent in [JSON format](https://en.wikipedia.org/wiki/JSON). In case the API request results in an error, it is represented by an `"errors": {}` key in the JSON response.
         3.  The request method (verb) determines the nature of action you intend to perform. A request made using the GET method implies that you want to fetch something from application, and POST implies you want to save something new to application.
         4.  **Info**, **message** and **devMessage** properties in JSON response are only available in _development mode._
         5.  The API calls will respond with appropriate [HTTP status codes](https://en.wikipedia.org/wiki/List_of_HTTP_status_codes) for all requests. Within Client, when a response is received, the status code is highlighted and is accompanied by a help text that indicates the possible meaning of the response code. A `200 OK` indicates all went well, while `4XX` or `5XX` response codes indicate an error from the requesting client or our API servers respectively.
             

...
             The rate limiting conditions are as follows:
...
             

         # Authentication

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
    /auth/refresh-token:
        post:
            tags:
                - 'auth'
            summary: 'Refresh token'
            description: 'Refresh token'
            operationId: 'refreshToken'
            requestBody:
                $ref: '#/components/requestBodies/RefreshToken'
            responses:
                '201':
                    $ref: '#/components/responses/RefreshTokenSuccess'
                '400':
                    $ref: '#/components/responses/BadRequestInCreate'
                '401':
                    description: 'Refresh token is expired'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiResponse'
                            examples:
                                REFRESH_TOKEN_EXPIRED:
                                    value:
                                        data: {}
                                        message: 'عدم احراز هویت'
                                        devMessage: 'refresh token is expired'
                                        msgCode: 40102
                                        info: {}
                '404':
                    # TODO: find better status code, 404 is not correct
                    $ref: '#/components/responses/UserNotFound'
                '406':
                    # TODO: Not Acceptable is not correct, gone (410) or conflict (409) is more suitable
                    description: 'Not Acceptable'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiResponse'
                            examples:
                                TOKEN_DOES_NOT_EXIST:
                                    value:
                                        data: {}
                                        message: 'دیتای ورودی معتبر نیست'
                                        devMessage: 'توکن وجود ندارد'
                                        msgCode: 40601
                                        info: {}
                '422':
                    $ref: '#/components/responses/ValidationError'
...
components:
    examples:
...
        JSON_SYNTAX_ERROR:
            description: 'JSON syntax error [read more](#JSON-Syntax-Error)'
            # TODO: add more details to Overview section
            value:
                message: 'JSON syntax error'
                devMessage: 'JSON syntax error'
                msgCode: 40001
                info: {}
        EMPTY_INPUT_BODY:
            description: 'the request body is a empty json [read more](#empty-input-body)'
            # TODO: add more details to Overview section
            value:
                message: 'دیتای ورودی معتبر نیست'
                devMessage: 'Input body is empty'
                msgCode: 40003
                info: {}
        INPUT_DATA_INVALID:
            value:
                errors:
                    field: name
                    type: stringMin
                    message: 'نام باید حداقل 1 کاراکتر باشد'
                message: 'دیتای ورودی معتبر نیست'
                msgCode: 42201
                info: {}
...
        USER_NOT_FOUND:
            value:
                message: 'کاربر یافت نشد'
                msgCode: 40401
                info: {}
...
    schemas:
        refreshToken:
            type: object
            properties:
                refreshToken:
                    type: string
                    description: 'refresh token'
                    minLength: 128
                    maxLength: 128
        accessToken:
            type: object
            properties:
                accessToken:
                    type: string
                    description: 'access token'
...
        ApiResponse:
            type: object
            properties:
                data:
                    type: object
                    nullable: true
                message:
                    type: string
                msgCode:
                    type: number
                info:
                    oneOf:
                        - type: string
                        - type: object
                    example: {}
        ApiInvalidInputDataResponse:
            type: object
            properties:
                errors:
                    type: array
                    items:
                        type: object
                        properties:
                            message:
                                type: string
                            type:
                                type: string
                                enum:
                                    - required
                                    - email
                                    - emailEmpty
                                    - emailMax
                                    - emailMin
                                    - equalField
                                    - length
                                    - max
                                    - min
                                    - number
                                    - numberNotEqual
                                    - numberPositive
                                    - objectID
                                    - objectStrict
                                    - pattern
                                    - string
                                    - stringEmpty
                                    - stringEnum
                                    - stringLength
                                    - stringMax
                                    - stringMin
                                    - url
                                    - DB_ERROR
                            field:
                                type: string
                        nullable: true
                message:
                    type: string
                msgCode:
                    type: number
                info:
                    oneOf:
                        - type: string
                        - type: object
                    example: {}
    requestBodies:
...
        RefreshToken:
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            refreshToken:
                                $ref: '#/components/schemas/refreshToken'
                        required:
                            - refreshToken
                        example:
                            refreshToken: '23a2a5110799a22ccdd5819c8f811ea8879d0f00fd65ae1f79365634a344309b7b70f79f6ac02920915251accd814043d874263c911feb984a8660412b56f836'
...
    responses:
...
        ValidationError:
            description: 'Validation Error'
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiInvalidInputDataResponse'
                    examples:
                        INPUT_DATA_INVALID:
                            $ref: '#/components/examples/INPUT_DATA_INVALID'
...
        BadRequestInCreate:
            description: 'Bad Request'
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiResponse'
                    examples:
                        JSON_SYNTAX_ERROR:
                            $ref: '#/components/examples/JSON_SYNTAX_ERROR'
                        EMPTY_INPUT_BODY:
                            $ref: '#/components/examples/EMPTY_INPUT_BODY'
...
        RefreshTokenSuccess:
            description: Success
            content:
                application/json:
                    schema:
                        allOf:
                            - $ref: '#/components/schemas/ApiResponse'
                            - type: object
                              properties:
                                  message:
                                      example: 'عملیات موفقیت آمیز بود'
                                  msgCode:
                                      example: 201
                                  data:
                                      allOf:
                                          - $ref: '#/components/schemas/refreshToken'
                                          - $ref: '#/components/schemas/accessToken'
                    example:
                        data:
                            accessToken: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'
                            refreshToken: '23a2a5110799a22ccdd5819c8f811ea8879d0f00fd65ae1f79365634a344309b7b70f79f6ac02920915251accd814043d874263c911feb984a8660412b56f836'
                        message: 'عملیات موفقیت آمیز بود'
                        msgCode: 201
                        info: {}
...
        UserNotFound:
            description: 'Not found'
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiResponse'
                    examples:
                        USER_NOT_FOUND:
                            $ref: '#/components/examples/USER_NOT_FOUND'
...
    securitySchemes:
        accessToken:
            type: apiKey
            description: you can get this token from /auth/login endpoint at the response header
            name: x-access-token
            in: header
...